requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
django-json-widget==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
envanter==1.2.0
future==0.18.3
Pillow==10.0.0
psycopg2==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
